move nmtensor_registry
