+++ nemo/utils/nmtensor_registry.py
